backend/requirements.txt
flask
flask-cors
langchain
chromadb
gitpython
pygments
unidiff
requests
python-dotenv
langchain-community 